+++ src/index.ts
@@ -2,3 +2,4 @@ export { Canvas } from "./Canvas";
 export { Brush } from "./Brush";
 export * from "./utils";
 export * from "./types";
+export * from "./modules";